refactor: separate video player
