+++ src/components/shared/video-player/index.js
@@ -1,0 +1,3 @@
+import VideoPlayer from './video-player';
+
+export default VideoPlayer;
